migrate some cluster configuration into cmd_clusters
